refactor: move batch commands (check-all, optimize-all, fix-ios-all) to gallery sub-command
