+++ photree/cli/app.py
@@ -9,6 +9,7 @@ from .album_cmd import album_app
 from .check_cmd import check_cmd
 from .demo_cmd import demo_app
 from .export_cmd import export_app
+from .gallery_cmd import gallery_app
 from .import_cmd import import_app
 
 
@@ -45,4 +46,5 @@ app.command("check")(check_cmd)
 app.add_typer(album_app)
 app.add_typer(demo_app)
 app.add_typer(export_app)
+app.add_typer(gallery_app)
 app.add_typer(import_app)
