+++ photree/cli/gallery_cmd.py
@@ -1,0 +1,672 @@
+"""CLI commands for the ``photree gallery`` sub-app.
+
+Gallery commands operate on multiple albums at once (batch operations).
+"""
+
+from __future__ import annotations
+
+import io
+from collections.abc import Callable
+from contextlib import redirect_stdout
+from pathlib import Path
+from typing import Annotated, Optional
+
+import typer
+
+from ..album import (
+    fixes as album_fixes,
+    naming as album_naming,
+    optimize as album_optimize,
+    output as album_output,
+    preflight as album_preflight,
+)
+from ..album.exif import check_exiftool_available
+from ..fsprotocol import (
+    LinkMode,
+    discover_contributors,
+    display_path,
+)
+from .album_cmd import (
+    _check_sips_or_exit,
+    _run_fix_ios,
+    _validate_fix_flags,
+)
+from .progress import BatchProgressBar
+
+gallery_app = typer.Typer(
+    name="gallery",
+    help="Batch operations on multiple albums.",
+    no_args_is_help=True,
+)
+
+
+def _resolve_check_batch_albums(
+    base_dir: Path | None,
+    album_dirs: list[Path] | None,
+) -> tuple[list[Path], Path | None]:
+    """Resolve album list for check commands (all album types).
+
+    Uses :func:`discover_albums` which detects iOS albums, ``.album``
+    sentinels, and leaf directories.
+    """
+    return _resolve_batch_albums_with(
+        base_dir, album_dirs, album_preflight.discover_albums
+    )
+
+
+def _resolve_batch_albums(
+    base_dir: Path | None,
+    album_dirs: list[Path] | None,
+) -> tuple[list[Path], Path | None]:
+    """Resolve album list for iOS-specific commands.
+
+    Uses :func:`discover_ios_albums` which only finds albums with an
+    ``ios/`` subdirectory.
+    """
+    return _resolve_batch_albums_with(
+        base_dir, album_dirs, album_preflight.discover_ios_albums
+    )
+
+
+def _resolve_batch_albums_with(
+    base_dir: Path | None,
+    album_dirs: list[Path] | None,
+    discover_fn: Callable[[Path], list[Path]],
+) -> tuple[list[Path], Path | None]:
+    """Resolve album list from mutually exclusive --dir / --album-dir options.
+
+    Returns ``(albums, display_base)`` where *display_base* is the base
+    directory when --dir was used (for relative display names), or ``None``
+    when --album-dir was used (display names are CWD-relative).
+    """
+    from rich.progress import Progress, SpinnerColumn, TextColumn
+
+    if base_dir is not None and album_dirs is not None:
+        typer.echo(
+            "--dir and --album-dir are mutually exclusive.",
+            err=True,
+        )
+        raise typer.Exit(code=1)
+
+    if album_dirs is not None:
+        return (album_dirs, None)
+
+    # --dir mode (explicit or default)
+    resolved_base = base_dir if base_dir is not None else Path(".").resolve()
+    with Progress(
+        SpinnerColumn(),
+        TextColumn("[progress.description]{task.description}"),
+        transient=True,
+    ) as progress:
+        progress.add_task("Resolving album list...", total=None)
+        albums = discover_fn(resolved_base)
+    return (albums, resolved_base)
+
+
+def _display_name(album_dir: Path, base_dir: Path | None, cwd: Path) -> str:
+    """Human-readable album name relative to *base_dir* or *cwd*."""
+    if base_dir is not None:
+        return str(album_dir.relative_to(base_dir))
+
+    return str(display_path(album_dir, cwd))
+
+
+@gallery_app.command("check")
+def check_cmd(
+    base_dir: Annotated[
+        Optional[Path],
+        typer.Option(
+            "--dir",
+            "-d",
+            help="Base directory to recursively scan for iOS albums.",
+            exists=True,
+            file_okay=False,
+            resolve_path=True,
+        ),
+    ] = None,
+    album_dirs: Annotated[
+        Optional[list[Path]],
+        typer.Option(
+            "--album-dir",
+            "-a",
+            help="Album directory to check (repeatable).",
+            exists=True,
+            file_okay=False,
+            resolve_path=True,
+        ),
+    ] = None,
+    checksum: Annotated[
+        bool,
+        typer.Option(
+            "--checksum/--no-checksum",
+            help="Enable/disable SHA-256 checksum verification (default: enabled).",
+        ),
+    ] = True,
+    fatal_warnings: Annotated[
+        bool,
+        typer.Option(
+            "--fatal-warnings",
+            "-W",
+            help="Treat informational warnings (e.g. missing sidecars) as errors.",
+        ),
+    ] = False,
+    check_naming: Annotated[
+        bool,
+        typer.Option(
+            "--check-naming/--no-check-naming",
+            help="Enable/disable album naming convention checks (default: enabled).",
+        ),
+    ] = True,
+    check_date_collisions: Annotated[
+        bool,
+        typer.Option(
+            "--check-date-collisions/--no-check-date-collisions",
+            help="Enable/disable cross-album date collision detection (default: enabled).",
+        ),
+    ] = True,
+) -> None:
+    """Check all albums under a directory or from an explicit list."""
+
+    cwd = Path.cwd()
+    albums, display_base = _resolve_check_batch_albums(base_dir, album_dirs)
+
+    # System checks (once)
+    sips_available = album_preflight.check_sips_available()
+    exiftool_available = check_exiftool_available()
+    typer.echo("System Checks:")
+    typer.echo(album_output.sips_check(sips_available))
+    typer.echo(album_output.exiftool_check(exiftool_available))
+    if not sips_available:
+        typer.echo("")
+        typer.echo(album_output.sips_troubleshoot(), err=True)
+        raise typer.Exit(code=1)
+
+    if not albums:
+        typer.echo("\nNo albums found.")
+        raise typer.Exit(code=0)
+
+    if display_base is not None:
+        typer.echo(f"\nFound {len(albums)} album(s).\n")
+    else:
+        typer.echo("")
+
+    # Check each album
+    progress = BatchProgressBar(
+        total=len(albums), description="Checking", done_description="check"
+    )
+    passed = 0
+    failed_albums: list[Path] = []
+
+    for album_dir in albums:
+        album_name = _display_name(album_dir, display_base, cwd)
+
+        progress.on_start(album_name)
+        result = album_preflight.run_album_check(
+            album_dir,
+            sips_available=sips_available,
+            exiftool_available=exiftool_available,
+            checksum=checksum,
+            check_naming_flag=check_naming,
+        )
+
+        album_ok = result.success and not (fatal_warnings and result.has_warnings)
+        if album_ok:
+            progress.on_end(album_name, success=True)
+            passed += 1
+        else:
+            progress.on_end(album_name, success=False)
+            failed_albums.append(album_dir)
+
+    progress.stop()
+
+    # Batch naming checks (date collisions across all albums)
+    if check_naming and check_date_collisions:
+        parsed_albums = [
+            (album.name, parsed)
+            for album in albums
+            if (parsed := album_naming.parse_album_name(album.name)) is not None
+        ]
+        batch_naming = album_naming.check_batch_date_collisions(parsed_albums)
+        typer.echo("")
+        typer.echo(album_output.format_batch_naming_issues(batch_naming))
+        if not batch_naming.success:
+            failed_albums.append(albums[0])  # ensure non-zero exit
+
+    # Summary
+    typer.echo(album_output.batch_check_summary(passed, len(failed_albums)))
+
+    if failed_albums:
+        typer.echo("\nTo investigate failures:", err=True)
+        for album_dir in sorted(set(failed_albums)):
+            typer.echo(
+                f'  photree album check --album-dir "{display_path(album_dir, cwd)}"',
+                err=True,
+            )
+        raise typer.Exit(code=1)
+
+
+@gallery_app.command("fix")
+def fix_cmd(
+    base_dir: Annotated[
+        Optional[Path],
+        typer.Option(
+            "--dir",
+            "-d",
+            help="Base directory to recursively scan for albums.",
+            exists=True,
+            file_okay=False,
+            resolve_path=True,
+        ),
+    ] = None,
+    album_dirs: Annotated[
+        Optional[list[Path]],
+        typer.Option(
+            "--album-dir",
+            "-a",
+            help="Album directory to fix (repeatable).",
+            exists=True,
+            file_okay=False,
+            resolve_path=True,
+        ),
+    ] = None,
+    refresh_jpeg: Annotated[
+        bool,
+        typer.Option(
+            "--refresh-jpeg",
+            help="Refresh {contributor}-jpg/ from {contributor}-img/ for all contributors.",
+        ),
+    ] = False,
+    dry_run: Annotated[
+        bool,
+        typer.Option(
+            "--dry-run",
+            "-n",
+            help="Print what would happen without modifying files.",
+        ),
+    ] = False,
+) -> None:
+    """Fix all albums under a directory or from an explicit list.
+
+    Works on all contributor types (iOS + plain). At least one fix flag
+    must be specified.
+    """
+
+    if not refresh_jpeg:
+        typer.echo(
+            "No fix specified. Run photree album fix-all --help for available fixes.",
+            err=True,
+        )
+        raise typer.Exit(code=1)
+
+    if refresh_jpeg:
+        _check_sips_or_exit()
+
+    cwd = Path.cwd()
+    albums, display_base = _resolve_check_batch_albums(base_dir, album_dirs)
+
+    if not albums:
+        typer.echo("\nNo albums found.")
+        raise typer.Exit(code=0)
+
+    if display_base is not None:
+        typer.echo(f"\nFound {len(albums)} album(s).\n")
+
+    progress = BatchProgressBar(
+        total=len(albums), description="Fixing", done_description="fix"
+    )
+    fixed = 0
+    failed_albums: list[Path] = []
+
+    for album_dir in albums:
+        album_name = _display_name(album_dir, display_base, cwd)
+        progress.on_start(album_name)
+
+        try:
+            contributors = discover_contributors(album_dir)
+            if refresh_jpeg:
+                for contrib in contributors:
+                    if (album_dir / contrib.img_dir).is_dir():
+                        album_fixes.refresh_jpeg(album_dir, contrib, dry_run=dry_run)
+            progress.on_end(album_name, success=True)
+            fixed += 1
+        except Exception:
+            progress.on_end(album_name, success=False)
+            failed_albums.append(album_dir)
+
+    progress.stop()
+
+    typer.echo(f"\nDone. {fixed} album(s) fixed, {len(failed_albums)} failed.")
+
+    if failed_albums:
+        typer.echo("\nFailed albums:", err=True)
+        for album_dir in failed_albums:
+            typer.echo(
+                f'  photree album fix --album-dir "{display_path(album_dir, cwd)}"',
+                err=True,
+            )
+        raise typer.Exit(code=1)
+
+
+@gallery_app.command("optimize")
+def optimize_cmd(
+    base_dir: Annotated[
+        Optional[Path],
+        typer.Option(
+            "--dir",
+            "-d",
+            help="Base directory to recursively scan for iOS albums.",
+            exists=True,
+            file_okay=False,
+            resolve_path=True,
+        ),
+    ] = None,
+    album_dirs: Annotated[
+        Optional[list[Path]],
+        typer.Option(
+            "--album-dir",
+            "-a",
+            help="Album directory to optimize (repeatable).",
+            exists=True,
+            file_okay=False,
+            resolve_path=True,
+        ),
+    ] = None,
+    link_mode: Annotated[
+        LinkMode,
+        typer.Option(
+            "--link-mode",
+            help="How to create main files: hardlink (default), symlink, or copy.",
+        ),
+    ] = LinkMode.HARDLINK,
+    check: Annotated[
+        bool,
+        typer.Option(
+            "--check/--no-check",
+            help="Run integrity checks before optimizing (default: enabled).",
+        ),
+    ] = True,
+    checksum: Annotated[
+        bool,
+        typer.Option(
+            "--checksum/--no-checksum",
+            help="Enable/disable SHA-256 checksum verification (default: enabled).",
+        ),
+    ] = True,
+    dry_run: Annotated[
+        bool,
+        typer.Option(
+            "--dry-run",
+            "-n",
+            help="Print what would happen without modifying files.",
+        ),
+    ] = False,
+) -> None:
+    """Optimize all iOS albums under a directory or from an explicit list.
+
+    Runs integrity checks on each album first (unless --no-check), then
+    replaces main-img/ and main-vid/ file copies with links.
+    """
+
+    cwd = Path.cwd()
+    albums, display_base = _resolve_batch_albums(base_dir, album_dirs)
+
+    sips_available = True
+    if check:
+        # System checks (once)
+        sips_available = album_preflight.check_sips_available()
+        typer.echo("System Checks:")
+        typer.echo(album_output.sips_check(sips_available))
+        if not sips_available:
+            typer.echo("")
+            typer.echo(album_output.sips_troubleshoot(), err=True)
+            raise typer.Exit(code=1)
+
+    if not albums:
+        typer.echo("\nNo iOS albums found.")
+        raise typer.Exit(code=0)
+
+    if display_base is not None:
+        typer.echo(f"\nFound {len(albums)} iOS album(s).\n")
+    else:
+        typer.echo("")
+
+    # Check and optimize each album
+    progress = BatchProgressBar(
+        total=len(albums), description="Optimizing", done_description="optimize"
+    )
+    optimized = 0
+    failed_albums: list[Path] = []
+
+    for album_dir in albums:
+        album_name = _display_name(album_dir, display_base, cwd)
+
+        progress.on_start(album_name)
+
+        if check:
+            check_result = album_preflight.run_album_check(
+                album_dir,
+                sips_available=sips_available,
+                exiftool_available=False,
+                checksum=checksum,
+                check_naming_flag=False,
+            )
+            if not check_result.success:
+                progress.on_end(album_name, success=False)
+                failed_albums.append(album_dir)
+                continue
+
+        album_optimize.optimize_album(album_dir, link_mode=link_mode, dry_run=dry_run)
+        progress.on_end(album_name, success=True)
+        optimized += 1
+
+    progress.stop()
+
+    # Summary
+    typer.echo(album_output.batch_optimize_summary(optimized, len(failed_albums)))
+
+    if failed_albums:
+        typer.echo("\nTo investigate failures:", err=True)
+        for album_dir in failed_albums:
+            typer.echo(
+                f'  photree album check --album-dir "{display_path(album_dir, cwd)}"',
+                err=True,
+            )
+        raise typer.Exit(code=1)
+
+
+@gallery_app.command("fix-ios")
+def fix_ios_cmd(
+    base_dir: Annotated[
+        Optional[Path],
+        typer.Option(
+            "--dir",
+            "-d",
+            help="Base directory to recursively scan for iOS albums.",
+            exists=True,
+            file_okay=False,
+            resolve_path=True,
+        ),
+    ] = None,
+    album_dirs: Annotated[
+        Optional[list[Path]],
+        typer.Option(
+            "--album-dir",
+            "-a",
+            help="iOS album directory to fix (repeatable).",
+            exists=True,
+            file_okay=False,
+            resolve_path=True,
+        ),
+    ] = None,
+    link_mode: Annotated[
+        LinkMode,
+        typer.Option(
+            "--link-mode",
+            help="How to create main files: hardlink (default), symlink, or copy.",
+        ),
+    ] = LinkMode.HARDLINK,
+    refresh_combined: Annotated[
+        bool,
+        typer.Option(
+            "--refresh-combined",
+            help="Rebuild main-img/ and main-vid/ from orig/edit, then regenerate main-jpg/.",
+        ),
+    ] = False,
+    refresh_jpeg: Annotated[
+        bool,
+        typer.Option(
+            "--refresh-jpeg",
+            help="Refresh main-jpg/ from main-img/ (re-convert all HEIC→JPEG).",
+        ),
+    ] = False,
+    rm_upstream: Annotated[
+        bool,
+        typer.Option(
+            "--rm-upstream",
+            help="Propagate deletions from browsing dirs (main-jpg, main-vid) to upstream dirs.",
+        ),
+    ] = False,
+    rm_orphan: Annotated[
+        bool,
+        typer.Option(
+            "--rm-orphan",
+            help="Delete edited and main files that have no corresponding orig file.",
+        ),
+    ] = False,
+    prefer_higher_quality_when_dups: Annotated[
+        bool,
+        typer.Option(
+            "--prefer-higher-quality-when-dups",
+            help="Delete lower-quality duplicates (e.g. JPG when DNG or HEIC exists for the same number).",
+        ),
+    ] = False,
+    rm_orphan_sidecar: Annotated[
+        bool,
+        typer.Option(
+            "--rm-orphan-sidecar",
+            help="Delete AAE sidecar files that have no matching media file.",
+        ),
+    ] = False,
+    rm_miscategorized: Annotated[
+        bool,
+        typer.Option(
+            "--rm-miscategorized",
+            help="Delete files in the wrong directory (edited in orig or vice versa).",
+        ),
+    ] = False,
+    rm_miscategorized_safe: Annotated[
+        bool,
+        typer.Option(
+            "--rm-miscategorized-safe",
+            help="Delete miscategorized files only if they already exist in the correct directory.",
+        ),
+    ] = False,
+    mv_miscategorized: Annotated[
+        bool,
+        typer.Option(
+            "--mv-miscategorized",
+            help="Move files in the wrong directory to the correct one.",
+        ),
+    ] = False,
+    dry_run: Annotated[
+        bool,
+        typer.Option(
+            "--dry-run",
+            "-n",
+            help="Print what would happen without modifying files.",
+        ),
+    ] = False,
+) -> None:
+    """Apply fix-ios to all iOS albums under a directory or from an explicit list.
+
+    Accepts the same fix flags as fix-ios. At least one fix flag must be specified.
+    """
+
+    cwd = Path.cwd()
+
+    _validate_fix_flags(
+        refresh_combined=refresh_combined,
+        refresh_jpeg=refresh_jpeg,
+        rm_upstream=rm_upstream,
+        rm_orphan=rm_orphan,
+        rm_orphan_sidecar=rm_orphan_sidecar,
+        prefer_higher_quality_when_dups=prefer_higher_quality_when_dups,
+        rm_miscategorized=rm_miscategorized,
+        rm_miscategorized_safe=rm_miscategorized_safe,
+        mv_miscategorized=mv_miscategorized,
+    )
+
+    albums, display_base = _resolve_batch_albums(base_dir, album_dirs)
+
+    if not albums:
+        typer.echo("No iOS albums found.")
+        raise typer.Exit(code=0)
+
+    if display_base is not None:
+        typer.echo(f"Found {len(albums)} iOS album(s).\n")
+
+    progress = BatchProgressBar(
+        total=len(albums), description="Fixing", done_description="fix-ios"
+    )
+    fixed = 0
+    failed_albums: list[Path] = []
+    album_reports: list[tuple[str, str]] = []
+
+    for album_dir in albums:
+        album_name = _display_name(album_dir, display_base, cwd)
+        progress.on_start(album_name)
+
+        buf = io.StringIO()
+        try:
+            with redirect_stdout(buf):
+                _run_fix_ios(
+                    album_dir,
+                    link_mode=link_mode,
+                    dry_run=dry_run,
+                    log_cwd=cwd,
+                    show_progress=False,
+                    refresh_combined=refresh_combined,
+                    refresh_jpeg=refresh_jpeg,
+                    rm_upstream=rm_upstream,
+                    rm_orphan=rm_orphan,
+                    rm_orphan_sidecar=rm_orphan_sidecar,
+                    prefer_higher_quality_when_dups=prefer_higher_quality_when_dups,
+                    rm_miscategorized=rm_miscategorized,
+                    rm_miscategorized_safe=rm_miscategorized_safe,
+                    mv_miscategorized=mv_miscategorized,
+                )
+            progress.on_end(album_name, success=True)
+            fixed += 1
+            captured = buf.getvalue()
+            if captured.strip():
+                album_reports.append((album_name, captured))
+        except Exception:
+            progress.on_end(album_name, success=False)
+            failed_albums.append(album_dir)
+
+    progress.stop()
+
+    # Print per-album action reports (e.g. dry-run details).
+    # Use color=True to preserve ANSI escapes captured from the fix run.
+    if album_reports:
+        typer.echo("")
+        for album_name, report in album_reports:
+            typer.echo(f"{album_name}:")
+            typer.echo(report, color=True)
+
+    typer.echo(album_output.batch_fix_ios_summary(fixed, len(failed_albums)))
+
+    if failed_albums:
+        typer.echo("\nFailed albums:", err=True)
+        for album_dir in failed_albums:
+            typer.echo(
+                f'  photree album fix-ios --album-dir "{display_path(album_dir, cwd)}"',
+                err=True,
+            )
+        raise typer.Exit(code=1)
+
+
+# Re-register the export batch command from export_cmd
+from .export_cmd import export_all_cmd  # noqa: E402
+
+gallery_app.command("export")(export_all_cmd)
